+++ android-gif-drawable/src/main/java/pl/droidsonroids/gif/GifTexImage2D.java
@@ -126,6 +126,7 @@ public class GifTexImage2D {
 	}
 
 	@Override
+	@SuppressWarnings("ThrowFromFinallyBlock")
 	protected final void finalize() throws Throwable {
 		try {
 			recycle();